count the ngram.

+++ sequential/include/hash_table.h
@@ -1,11 +1,9 @@
 #ifndef HASH_TABLE_H
 #define HASH_TABLE_H
 
-#include <stdlib.h>
-
 typedef struct HashTable HashTable;
 
-HashTable *create_hash_table(size_t buckets_size);
+HashTable *create_hash_table(int buckets_size);
 void add_gram(HashTable *table, const char *gram);
 void free_hash_table(HashTable *table);
